pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vulnq"
version = "1.0.0"
description = "Lightweight vulnerability query tool for multi-source security data aggregation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Oscar Valenzuela B.", email = "oscar.valenzuela.b@gmail.com"}
]
maintainers = [
    {name = "Oscar Valenzuela B.", email = "oscar.valenzuela.b@gmail.com"}
]
keywords = [
    "vulnerability",
    "security",
    "cve",
    "osv",
    "nvd",
    "purl",
    "cpe",
    "vulnerability-scanner",
    "vulnerability-database",
    "sbom",
    "sca",
    "supply-chain"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Software Distribution",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "packageurl-python>=0.11.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.11.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0"
]
cache = [
    "diskcache>=5.0.0"
]
all = [
    "vulnq[dev,cache]"
]

[project.urls]
Homepage = "https://github.com/scanoss/vulnq"
Documentation = "https://github.com/scanoss/vulnq/blob/main/README.md"
Repository = "https://github.com/scanoss/vulnq.git"
Issues = "https://github.com/scanoss/vulnq/issues"
Changelog = "https://github.com/scanoss/vulnq/blob/main/CHANGELOG.md"

[project.scripts]
vulnq = "vulnq.cli:main"

[tool.setuptools]
packages = ["vulnq"]

[tool.setuptools.package-data]
vulnq = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["vulnq"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/test_*.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass"
]